+++ virtualargofleet/__init__.py
@@ -1,5 +1,5 @@
-from .virtualargofleet import VelocityField as velocityfield
 from .virtualargofleet import VirtualFleet as virtualfleet
+from .velocity_helpers import VelocityField as velocityfield
 
 #
 __all__ = (
